Revert "refacto: Create component Menu"
This reverts commit a6d6b9136bf977ee89feec772aea392eae60df9d.
